pipelines/requirements.txt
apache-beam==2.25.0
appdirs==1.4.4
avro-python3==1.9.2.1
black==20.8b1
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.1.0
fasteners==0.15
future==0.18.2
google-api-core==1.23.0
google-apitools==0.5.31
google-auth==1.23.0
google-cloud-bigquery==1.28.0
google-cloud-bigtable==1.5.1
google-cloud-build==2.0.0
google-cloud-core==1.4.3
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-spanner==1.19.1
google-cloud-videointelligence==1.16.0
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.33.2
grpcio-gcp==0.2.2
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
libcst==0.3.13
mock==2.0.0
monotonic==1.5
mypy-extensions==0.4.3
numpy==1.19.4
oauth2client==4.1.3
pathspec==0.8.1
pbr==5.5.1
proto-plus==1.11.0
protobuf==3.14.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
rsa==4.6
six==1.15.0
slack-sdk==3.0.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
